Handle EMR Exceptions in FlintCancelJob Operation
