+++ spark/src/test/java/org/opensearch/sql/spark/execution/session/InteractiveSessionTest.java
@@ -227,7 +227,8 @@ public class InteractiveSessionTest extends OpenSearchIntegTestCase {
     }
 
     @Override
-    public CancelJobRunResult cancelJobRun(String applicationId, String jobId) {
+    public CancelJobRunResult cancelJobRun(
+        String applicationId, String jobId, boolean allowExceptionPropagation) {
       cancelJobRunCalled++;
       return null;
     }
